fix pgpool tests to not clobber other tests
The tests are all run in the same process against the same instances of
the module so care must be taken to not clobber things like the shared
config that the tests all use.

When the username is removed from the shared config then the pg attempts
to use the OS username as the DB username. This is not guaranteed and
is actually not true in the Windows environments these tests run under
in CI.