Extended Operator : Arrays overlap
The 'contains' keyword search if the value contains ALL the values in the submitted array. I need a function to find a row which contains any of values in the submitted array.
I propose to support the && array operator: describe as is in the official postgresql documentation : "Do the arrays overlap, that is, have any elements in common?"
https://www.postgresql.org/docs/current/functions-array.html
The new name of this function could be "overlaps".